Refactoring the drop_counters test to reuse existing methods
