+++ tests/common/helpers/drop_counters/combined_drop_counters.yml
